+++ packages/TokenRunner/src/ReflectionDocBlock/CleanDocBlockFactory.php
@@ -6,11 +6,11 @@ use phpDocumentor\Reflection\DocBlock;
 use phpDocumentor\Reflection\DocBlock\DescriptionFactory;
 use phpDocumentor\Reflection\DocBlock\StandardTagFactory;
 use phpDocumentor\Reflection\DocBlock\Tags\Param;
-use phpDocumentor\Reflection\DocBlock\Tags\Return_;
 use phpDocumentor\Reflection\DocBlock\Tags\Var_;
 use phpDocumentor\Reflection\DocBlockFactory;
 use phpDocumentor\Reflection\FqsenResolver;
 use phpDocumentor\Reflection\TypeResolver;
+use Symplify\TokenRunner\ReflectionDocBlock\Tag\TolerantParam;
 use Symplify\TokenRunner\ReflectionDocBlock\Tag\TolerantReturn;
 
 /**
@@ -27,7 +27,7 @@ final class CleanDocBlockFactory
     {
         $fqsenResolver = new FqsenResolver();
         $tagFactory = new StandardTagFactory($fqsenResolver, [
-            'param' => Param::class,
+            'param' => TolerantParam::class,
             'return' => TolerantReturn::class,
             'var' => Var_::class,
         ]);