+++ packages/TokenRunner/src/Wrapper/FixerWrapper/DocBlockWrapper.php
@@ -18,6 +18,7 @@ use Symplify\TokenRunner\DocBlock\ArrayResolver;
 use Symplify\TokenRunner\Guard\TokenTypeGuard;
 use Symplify\TokenRunner\ReflectionDocBlock\CleanDocBlockFactory;
 use Symplify\TokenRunner\ReflectionDocBlock\DocBlockSerializerFactory;
+use Symplify\TokenRunner\ReflectionDocBlock\Tag\TolerantParam;
 
 final class DocBlockWrapper
 {
@@ -275,11 +276,11 @@ final class DocBlockWrapper
         return ltrim(trim($content), '\\');
     }
 
-    private function findParamTagByName(string $name): ?Param
+    private function findParamTagByName(string $name): ?TolerantParam
     {
         $paramTags = $this->phpDocumentorDocBlock->getTagsByName('param');
 
-        /** @var Param $paramTag */
+        /** @var TolerantParam $paramTag */
         foreach ($paramTags as $paramTag) {
             if ($paramTag->getVariableName() === $name) {
                 return $paramTag;