+++ CHANGELOG.md
@@ -7,14 +7,14 @@
 
 ### Changed
 
-- [#508] `RemoveUselessDocBlockFixer` is now configurable to accept types with your own preferences
+- [#508] `RemoveUselessDocBlockFixer` is now configurable to accept types to remove with your own preferences
 
     ```yaml
     # easy-coding-standard.neon
     checkers:
         Symplify\CodingStandard\Fixer\Commenting\RemoveUselessDocBlockFixer:
-            # defaults
-            useful_types: ['mixed', 'object']
+            useless_types: ['mixed']
+            # "[]" is default 
     ```
 
 ### Added